requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
faiss-cpu==1.7.4
flask==2.3.3
markdown==3.4.4
spacy==3.6.1
gensim==4.3.1
tqdm==4.66.1
